+++ app/formSchema/pages/additionalProjectInformation.ts
@@ -1,23 +1,32 @@
-const additionalProjectInformation = {
-  additionalProjectInformation: {
+const projectInformation = {
+  projectInformation: {
     title: 'Project information',
+    description:
+      'PLEASE NOTE: If the project is approved, the project information herein may be used, in whole or in part, in publicly accessible websites, media releases, or other similar material.',
     type: 'object',
     properties: {
-      projectTitle: {
-        title: 'Project title',
-        type: 'string',
+      projectSpanMultipleLocations: {
+        title: 'Does your project span multiple provinces/territories?',
+        type: 'boolean',
+        enum: [true, false],
+        enumNames: ['Yes', 'No'],
       },
-      geographicAreaDescription: {
-        title: 'Geographic project area description',
-        type: 'string',
-      },
-      projectDescription: {
-        title:
-          'Using non-technical language, provide a description of the project, its key elements, objectives, and benefits',
-        type: 'string',
+      projectLocations: {
+        title: 'If yes, province or territory location (check all that apply)',
+        type: 'array',
+        items: {
+          type: 'string',
+          enum: [
+            'Alberta',
+            'British Columbia',
+            'Northwest Territories',
+            'Yukon',
+          ],
+        },
+        uniqueItems: true,
       },
     },
   },
 };
 
-export default additionalProjectInformation;
+export default projectInformation;
