+++ app/formSchema/pages/index.ts
@@ -1,9 +1,9 @@
 export { default as alternateContact } from './alternateContact';
-export { default as additionalProjectInformation } from './additionalProjectInformation';
+export { default as additionalProjectInformation } from './projectInformation';
 export { default as authorizedContact } from './authorizedContact';
 export { default as budgetDetails } from './budgetDetails';
 export { default as contactInformation } from './contactInformation';
 export { default as existingNetworkCoverage } from './existingNetworkCoverage';
 export { default as organizationLocation } from './organizationLocation';
 export { default as organizationProfile } from './organizationProfile';
-export { default as projectInformation } from './projectInformation';
+export { default as projectInformation } from './additionalProjectInformation';
