+++ app/formSchema/pages/projectInformation.ts
@@ -1,32 +1,21 @@
-const projectInformation = {
-  projectInformation: {
+const additionalProjectInformation = {
+  additionalProjectInformation: {
     title: 'Project information',
-    description:
-      'PLEASE NOTE: If the project is approved, the project information herein may be used, in whole or in part, in publicly accessible websites, media releases, or other similar material.',
     type: 'object',
+    description:
+      'For successful applicants receiving funding under CCBC, the contribution amounts will be made public via government websites, media releases or other similar materials. Information provided may be used, in whole or in part to assist the Province of BC and ISED in the preparation of any public announcements.',
     properties: {
-      projectSpanMultipleLocations: {
-        title: 'Does your project span multiple provinces/territories?',
-        type: 'boolean',
-        enum: [true, false],
-        enumNames: ['Yes', 'No'],
+      projectTitle: {
+        type: 'string',
       },
-      projectLocations: {
-        title: 'If yes, province or territory location (check all that apply)',
-        type: 'array',
-        items: {
-          type: 'string',
-          enum: [
-            'Alberta',
-            'British Columbia',
-            'Northwest Territories',
-            'Yukon',
-          ],
-        },
-        uniqueItems: true,
+      geographicAreaDescription: {
+        type: 'string',
+      },
+      projectDescription: {
+        type: 'string',
       },
     },
   },
 };
 
-export default projectInformation;
+export default additionalProjectInformation;
