+++ app/formSchema/uiSchema.ts
@@ -67,6 +67,33 @@ const uiSchema = {
     'totalProjectCost',
     'requestedCCBCFunding',
   ],
+  projectTitle: {
+    'ui:description': 'maximum 200 characters',
+    'ui:title':
+      'Project title. Be descriptive about the geographic region. We advise not using years in the title.',
+    'ui:options': {
+      maxLength: MAX_LONG_INPUT_LENGTH,
+    },
+  },
+  geographicAreaDescription: {
+    'ui:description': 'maximum 200 characters',
+    'ui:title':
+      'Geographic project area description. Describe the geographic location of the project area (i.e., include the closest communities and the general area which the project will target).',
+
+    'ui:options': {
+      maxLength: MAX_LONG_INPUT_LENGTH,
+    },
+  },
+  projectDescription: {
+    'ui:widget': 'TextAreaWidget',
+    'ui:description': 'maximum 3,500 characters',
+    'ui:title':
+      'Using non-technical language, provide a description of the project, including its key elements, purpose, objectives, and benefits. Identify the ‘who’, ‘what’, ‘where’, ‘when’, and ‘why’. Please avoid including confidential or proprietary information.',
+
+    'ui:options': {
+      maxLength: MAX_TEXTAREA_LENGTH,
+    },
+  },
   organizationName: {
     'ui:options': {
       maxLength: MAX_LONG_INPUT_LENGTH,
